Bugfix: changing range in get_next function from uint32_t to uint64_t
Due to overflow, if we call get_next function with 0, UINT32_T_MAX
arguments, we get SIGFPE due to division by 0. Changing type of modulo
from uint32_t to uint64_t prevents the overflow.

+++ mt_rng.cpp
@@ -47,7 +47,7 @@ uint32_t get_next(uint32_t min, uint32_t max)
     {
         return min;
     }
-    uint32_t range = (uint32_t)(max - min + 1);
+    uint64_t range = (uint64_t)(max - min) + 1;
     uint32_t random_value = mersenne_twister();
     uint32_t scaled_random = random_value % range;
     return min + scaled_random;
